decouple utils function
