+++ facex/core.py
@@ -1,8 +1,8 @@
 
-from skimage.transform import resize
 import numpy as np
 import pathlib
 import os
+from facex.utils import _preprocess_img
 
 weight = os.path.join(pathlib.Path(__file__).parent.absolute(), "model_optimized.tflite")
 
@@ -19,12 +19,7 @@ output_details = interpreter.get_output_details()
 
 
 
-def _preprocess_img(raw):
-    img = resize(raw,(200,200, 3))
-    img = np.expand_dims(img,axis=0)
-    if(np.max(img)>1):
-        img = img/255.0
-    return img
+
 
 def _parse_result(res):
     cat = ['anger', 'disgust', 'fear', 'happy', 'neutral', 'sadness', 'surprised']
